Make some progress towards reproducible builds on Windows
The patch to Makefile.am causes "make dist" to create the file
source_date_epoch containing the last git commit timestamp (even if that
file already exists), and to include that file in the tarball.

Then INSTALL.mingw uses the contents of that file to set the
$SOURCE_DATE_EPOCH environment variable, and there's a new patch to
libgcrypt (from Debian) that uses that environment variable to
override the build timestamp, which otherwise gets baked into the
binary's version string.