More towards reproducible builds on Windows
Preserve file timestamps when patching libgcrypt, since when that
package regenerates configure from configure.ac, it needs to look
in the libgcrypt git repository for versioning information.  Therefore
we manually patch both configure and configure.ac, and don't update the
file timestamps so that make doesn't try to rebuild configure.

Also use a deterministic order of files when creating the zip file.