Less nondeterminism in the zip file
We don't need to run zip under faketime if we use the -o option
(which sets the timestamp for the archive itself to equal the oldest
timestamp of a file in the archive).  We also add the -X option, which
removes Unix uid/gids and extended timestamps.  We also use
chmod -R u=rwX,go=rX to canonicalize the permissions on the files, which
are stored in the zip file.